removed client side code from notification cloud function
